requirements.txt
# ChessTutorAI - Chess Lesson Generation Engine
# Core dependencies for lesson generation

# Chess library for position analysis and move validation
python-chess>=1.999

# AI/ML providers for content generation
openai>=1.0.0

# HTTP client for API requests
httpx>=0.24.0
requests>=2.28.0

# Data compression for Lichess puzzle database
zstandard>=0.19.0

# Database and data handling
sqlite3  # Built-in with Python

# Optional: Enhanced logging and utilities
colorama>=0.4.6  # For colored terminal output
tqdm>=4.64.0     # For progress bars during imports

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0